+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 3.20.10 (May 13, 2024)
+
+### Removed
+
+- Remove `removeClasses()` util function
+
+### Fixed
+
+- Fix slide transition in Slider component in Safari
+
 ## 3.20.9 (May 8, 2024)
 
 ### Added